Django_Ml/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
filelock==3.18.0
gdown==5.2.0
idna==3.10
joblib==1.1.0
numpy==2.2.4
packaging==25.0
pandas==2.2.3
py4j==0.10.9.7
pyproject_hooks==1.2.0
PySocks==1.7.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
